DNS test file renames and refactorings
